+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-version = 0.1.0
+version = 0.2.0
 description = Recipes for approximate nearest neighbord search
 url = https://github.com/explosion/prodigy-ann
 author = Explosion
@@ -14,11 +14,11 @@ install_requires =
 
 [options.entry_points]
 prodigy_recipes =
-    ann.text.index = prodigy_ann:index
-    ann.text.fetch = prodigy_ann:fetch
-    textcat.ann.manual = prodigy_ann:textcat_ann_manual
-    ner.ann.manual = prodigy_ann:ner_ann_manual
-    spans.ann.manual = prodigy_ann:spans_ann_manual
+    ann.text.index = prodigy_ann.text:text_index
+    ann.text.fetch = prodigy_ann.text:text_fetch
+    textcat.ann.manual = prodigy_ann.text:textcat_ann_manual
+    ner.ann.manual = prodigy_ann.text:ner_ann_manual
+    spans.ann.manual = prodigy_ann.text:spans_ann_manual
 
 [bdist_wheel]
 universal = true
